refactor: enum class Gender 패키지 변경 및 내부 필드명 수정
