+++ BE/src/main/java/ifive/idrop/child/domain/Child.java
@@ -1,7 +1,7 @@
 package ifive.idrop.child.domain;
 
 import ifive.idrop.entity.PickUpInfo;
-import ifive.idrop.entity.enums.Gender;
+import ifive.idrop.common.enums.Gender;
 import ifive.idrop.parent.domain.Parent;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -10,19 +10,16 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-@Getter
 @Entity
+@Getter
 public class Child {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "child_id")
     private Long id;
     private String name;
-    private LocalDate birth;
-
+    private LocalDate birthDate;
     @Enumerated(EnumType.STRING)
     private Gender gender;
-
     private String image;
 
     @ManyToOne(fetch = FetchType.LAZY)
