+++ BE/src/main/java/ifive/idrop/driver/dto/DriverSubscribeInfoResponse.java
@@ -42,8 +42,8 @@ public class DriverSubscribeInfoResponse {
                 .parentName(parent.getName())
                 .parentPhoneNumber(parent.getPhoneNumber())
                 .childName(child.getName())
-                .childBirth(child.getBirth())
-                .childGender(child.getGender().getGender())
+                .childBirth(child.getBirthDate())
+                .childGender(child.getGender().getLabel())
                 .childImage(child.getImage())
                 .startDate(startDate)
                 .endDate(endDate)
